Add SecurityAnalysisResultRepository
