+++ src/main/java/org/gridsuite/securityanalysis/server/service/SecurityAnalysisService.java
@@ -6,21 +6,15 @@
  */
 package org.gridsuite.securityanalysis.server.service;
 
-import com.powsybl.contingency.BranchContingency;
-import com.powsybl.contingency.Contingency;
-import com.powsybl.contingency.ContingencyElement;
-import com.powsybl.contingency.GeneratorContingency;
-import com.powsybl.security.*;
-import org.gridsuite.securityanalysis.server.repository.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.powsybl.security.LimitViolationType;
+import com.powsybl.security.SecurityAnalysisResult;
+import org.gridsuite.securityanalysis.server.repository.SecurityAnalysisResultRepository;
 import org.springframework.stereotype.Service;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import java.util.Objects;
+import java.util.Set;
+import java.util.UUID;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
@@ -28,21 +22,12 @@ import java.util.stream.Stream;
 @Service
 public class SecurityAnalysisService {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(SecurityAnalysisService.class);
-
-    private ComputationStatusRepository computationStatusRepository;
-
-    private ContingencyRepository contingencyRepository;
-
-    private LimitViolationRepository limitViolationRepository;
+    private SecurityAnalysisResultRepository resultRepository;
 
     private SecurityAnalysisRunPublisherService runPublisherService;
 
-    public SecurityAnalysisService(ComputationStatusRepository computationStatusRepository, ContingencyRepository contingencyRepository,
-                                   LimitViolationRepository limitViolationRepository, SecurityAnalysisRunPublisherService runPublisherService) {
-        this.computationStatusRepository = Objects.requireNonNull(computationStatusRepository);
-        this.contingencyRepository = Objects.requireNonNull(contingencyRepository);
-        this.limitViolationRepository = Objects.requireNonNull(limitViolationRepository);
+    public SecurityAnalysisService(SecurityAnalysisResultRepository resultRepository, SecurityAnalysisRunPublisherService runPublisherService) {
+        this.resultRepository = Objects.requireNonNull(resultRepository);
         this.runPublisherService = Objects.requireNonNull(runPublisherService);
     }
 
@@ -57,78 +42,15 @@ public class SecurityAnalysisService {
         return Mono.just(resultUuid);
     }
 
-    private static LimitViolation fromEntity(LimitViolationEntity entity) {
-        return new LimitViolation(entity.getSubjectId(), entity.getLimitType(), entity.getLimitName(), entity.getAcceptableDuration(),
-                entity.getLimit(), entity.getLimitReduction(), entity.getValue(), entity.getSide());
-    }
-
-    private static boolean fromEntity(ComputationStatusEntity entity) {
-        return entity.isOk();
-    }
-
-    private static Contingency fromEntity(ContingencyEntity entity) {
-        Stream<ContingencyElement> branchStream = entity.getBranchIds() != null
-                ? entity.getBranchIds().stream().map(BranchContingency::new)
-                : Stream.empty();
-        Stream<ContingencyElement> generatorStream = entity.getGeneratorIds() != null
-                ? entity.getGeneratorIds().stream().map(GeneratorContingency::new)
-                : Stream.empty();
-        return new Contingency(entity.getContingencyId(), Stream.concat(branchStream, generatorStream).collect(Collectors.toList()));
-    }
-
     public Mono<SecurityAnalysisResult> getResult(UUID resultUuid, Set<LimitViolationType> limitTypes) {
-        Objects.requireNonNull(resultUuid);
-        Objects.requireNonNull(limitTypes);
-
-        Mono<Map<String, Boolean>> computationsStatusesMono = computationStatusRepository.findByResultUuid(resultUuid)
-                .collectMap(ComputationStatusEntity::getContingencyId, SecurityAnalysisService::fromEntity);
-        Mono<Map<String, Collection<LimitViolation>>> limitViolationsMono
-                = (limitTypes.isEmpty() ? limitViolationRepository.findByResultUuid(resultUuid) : limitViolationRepository.findByResultUuidAndLimitTypeIn(resultUuid, limitTypes))
-                .collectMultimap(LimitViolationEntity::getContingencyId, SecurityAnalysisService::fromEntity);
-        Mono<List<Contingency>> contingenciesMono = contingencyRepository.findByResultUuid(resultUuid)
-                .map(SecurityAnalysisService::fromEntity)
-                .collectList();
-
-        return Mono.zip(computationsStatusesMono, limitViolationsMono, contingenciesMono)
-                .flatMap(tuple -> {
-                    Map<String, Boolean> computationsStatuses = tuple.getT1();
-                    Map<String, Collection<LimitViolation>> limitViolations = tuple.getT2();
-                    List<Contingency> contingencies = tuple.getT3();
-
-                    if (computationsStatuses.isEmpty()) {
-                        return Mono.empty();
-                    }
-
-                    LimitViolationsResult preContingencyResult = new LimitViolationsResult(computationsStatuses.get(""),
-                            new ArrayList<>(limitViolations.getOrDefault("", Collections.emptyList())));
-
-                    List<PostContingencyResult> postContingencyResults = contingencies.stream()
-                            .map(contingency -> new PostContingencyResult(contingency,
-                                    computationsStatuses.get(contingency.getId()),
-                                    new ArrayList<>(limitViolations.getOrDefault(contingency.getId(), Collections.emptyList()))))
-                            .collect(Collectors.toList());
-
-                    return Mono.just(new SecurityAnalysisResult(preContingencyResult, postContingencyResults));
-                })
-                .doOnError(throwable -> LOGGER.error(throwable.toString(), throwable));
+        return resultRepository.find(resultUuid, limitTypes);
     }
 
     public Mono<Void> deleteResult(UUID resultUuid) {
-        Objects.requireNonNull(resultUuid);
-        Mono<Void> v1 = computationStatusRepository.deleteByResultUuid(resultUuid);
-        Mono<Void> v2 = limitViolationRepository.deleteByResultUuid(resultUuid);
-        Mono<Void> v3 = contingencyRepository.deleteByResultUuid(resultUuid);
-        return Flux.concat(v1, v2, v3)
-                .then()
-                .doOnError(throwable -> LOGGER.error(throwable.toString(), throwable));
+        return resultRepository.delete(resultUuid);
     }
 
     public Mono<Void> deleteResults() {
-        Mono<Void> v1 = computationStatusRepository.deleteAll();
-        Mono<Void> v2 = limitViolationRepository.deleteAll();
-        Mono<Void> v3 = contingencyRepository.deleteAll();
-        return Flux.concat(v1, v2, v3)
-                .then()
-                .doOnError(throwable -> LOGGER.error(throwable.toString(), throwable));
+        return resultRepository.deleteAll();
     }
 }
